Before changing of deckhub

+++ app/handlers/deckhub.py
@@ -5,13 +5,19 @@ from aiogram.enums import ParseMode
 from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import StatesGroup, State
-from aiogram.types import Message, CallbackQuery, ReplyKeyboardRemove
-
+from aiogram.methods import SetChatMenuButton
+from aiogram.types import Message, CallbackQuery, ReplyKeyboardRemove, BotCommand, MenuButtonCommands, \
+    BotCommandScopeDefault, BotCommandScopeChat
+from bot import bot
 import app.keyboards as kb
 from app.handlers import card_mode_start
 from app.requests import send_request
-from app.services import generate_deck_list_text
+from app.services import generate_deck_list_text, check_current_state, display_message_and_redirect
 from settings import BASE_URL
+
+from aiogram import types
+from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
+
 
 router = Router()
 
@@ -20,11 +26,17 @@ class DeckRename(StatesGroup):
     new_name = State()
 
 
+class DeckDelete(StatesGroup):
+    active = State()
+
+
 class ImportCards(StatesGroup):
     data = State()
 
+
 class DeckViewingState(StatesGroup):
     active = State()
+
 
 class CardManage(StatesGroup):
     card_ops_state = State()
@@ -35,19 +47,21 @@ class CardManage(StatesGroup):
     is_two_sides = State()
 
 
-
 @router.message(F.text.in_(('Decks', 'Back to desks list')))
 @router.message(Command('to_desks_list'))
 async def decks_list(message: Message, state: FSMContext):
     tg_id = state.key.user_id
     get_decks_url = f'{BASE_URL}/deck/api/v1/manage/{tg_id}/'
-    decks_data = await send_request(get_decks_url)
-    if decks_data:
+    response = await send_request(get_decks_url)
+    if response.get('status') // 100 == 2:
         await state.set_state(DeckViewingState.active)
-        await message.answer(f'🗃 *{f"__Decks__":^50}* 🗃', parse_mode=ParseMode.MARKDOWN_V2)
-        for deck in decks_data:
+        await message.answer(f'🗃 *{f"__Decks__":^50}* 🗃', parse_mode=ParseMode.MARKDOWN_V2,
+                             reply_markup=await kb.deckhub_manage_button())
+        for deck in response.get('data'):
             text = await generate_deck_list_text(deck)
             await message.answer(text, parse_mode=ParseMode.MARKDOWN_V2, reply_markup=await kb.decks_list(deck))
+    else:
+        await message.answer('Something went wrong. Please press REFRESH button.', reply_markup=kb.refresh_session)
 
 
 @router.callback_query(F.data == 'to_decks_list')
@@ -65,12 +79,14 @@ async def launch_card_mode(callback: CallbackQuery, state: FSMContext):
 
 
 @router.callback_query(F.data.startswith('manage_deck_'))
+@check_current_state
 async def manage_deck(callback: CallbackQuery, state: FSMContext):
     slug = callback.data.split('_')[-1]
     await callback.message.edit_reply_markup(reply_markup=await kb.manage_deck(slug))
 
 
 @router.callback_query(F.data.startswith('rename_deck_'))
+@check_current_state
 async def rename_deck(callback: CallbackQuery, state: FSMContext):
     slug = callback.data.split('_')[-1]
     await state.set_state(DeckRename.new_name)
@@ -85,37 +101,51 @@ async def rename_deck(callback: CallbackQuery, state: FSMContext):
 async def rename_deck_handler(message: Message, state: FSMContext):
     data = await state.get_data()
     new_name = message.text
-    if new_name and new_name.isalnum():
+    if len(new_name) > 100:
+        text = '❕️ New name must contain maximum 100 chars.'
+        return await display_message_and_redirect(message, state, text)
+
+    if new_name and any(char.isalnum() for char in new_name):
         new_name = new_name.capitalize().strip()
         slug = data.get('slug')
         url = f'{BASE_URL}/deck/api/v1/manage/{slug}/'
-        request = await send_request(url, method='PUT', data={'name': new_name})
-        if request and request.get('name', '').capitalize() == new_name:
-            await message.reply('Deck name was successfully changed.')
-            await asyncio.sleep(1)
-        await state.clear()
-        await decks_list(message, state)
+        response = await send_request(url, method='PUT', data={'name': new_name})
+        status = response.get('status', 0)
+        if status // 100 == 2:
+            text = 'Deck name was successfully changed.'
+        else:
+            text = 'Something went wrong.'
     else:
-        await message.answer('Name must consist of letters or numbers. Please try again.')
-        await asyncio.sleep(2)
-        await state.clear()
-        await decks_list(message, state)
+        text = 'Name must consist of letters or numbers. Please try again.'
+    await display_message_and_redirect(message, state, text)
 
 
 @router.callback_query(F.data.startswith('delete_deck_'))
-async def deck_delete(callback: CallbackQuery, state: FSMContext):
-    slug = callback.data.split('_')[-1]
-    url = f'{BASE_URL}/deck/api/v1/manage/{slug}/'
-    request = await send_request(url, method='DELETE')
-    if request == 204:
-        text = 'Deck was successfully deleted.'
-    else:
-        text = 'Something went wrong.'
+@check_current_state
+async def confirm_delete_deck(callback: CallbackQuery, state: FSMContext):
     await callback.answer()
-    await callback.message.answer(text, reply_markup=kb.studying_start)
-    await asyncio.sleep(1)
-    await decks_list(callback.message, state)
+    await state.set_state(DeckDelete.active)
+    slug = callback.data.split('_')[-1]
+    text = 'Are you sure to delete this deck?'
+    await callback.message.answer(text, reply_markup=await kb.confirm_delete_desk(slug))
 
+
+@router.callback_query(F.data.startswith('confirm_delete_deck_'))
+async def deck_delete(callback: CallbackQuery, state: FSMContext):
+    current_state = await state.get_state()
+    if current_state == DeckDelete.active:
+        slug = callback.data.split('_')[-1]
+        url = f'{BASE_URL}/deck/api/v1/manage/{slug}/'
+        response = await send_request(url, method='DELETE')
+        if response.get('status') == 204:
+            text = '🧨 Deck was successfully deleted.'
+            await callback.message.delete()
+        else:
+            text = 'Something went wrong.'
+            await callback.message.delete_reply_markup()
+        await callback.message.answer(text, reply_markup=kb.studying_start)
+        await asyncio.sleep(1.5)
+    await decks_list(callback.message, state)
 
 # @router.message(AddCard.front_side)
 # @router.message(AddCard.back_side)
@@ -138,3 +168,31 @@ async def deck_delete(callback: CallbackQuery, state: FSMContext):
 #         await state.clear()
 #         await decks_list(message, state)
 
+# @router.callback_query(F.data == 'deckhub_manage')
+# async def deckhub_manage(callback: CallbackQuery, state: FSMContext):
+#     await callback.message.edit_reply_markup()
+
+
+@router.message(F.text == 'b')
+async def send_welcome(message: types.Message):
+    user_commands = [
+        BotCommand(command='for', description='Описание команды /for'),
+        BotCommand(command='for2', description='Описание команды /for2'),
+        BotCommand(command='magik', description='Описание команды /magik'),
+        BotCommand(command='magik22', description='Описание команды /magik'),
+        BotCommand(command='magik2ww2', description='Описание команды /magik'),
+    ]
+
+    # Предполагаем, что у вас используется aiogram
+    menu_button = MenuButtonCommands(commands=user_commands)
+
+    res = await bot.set_my_commands(user_commands, scope=BotCommandScopeChat(chat_id=message.chat.id))
+    # await bot(SetChatMenuButton(menu_button=menu_button))
+
+# @router.message(F.text == 'a')
+# async def send_welcome(message: types.Message):
+#
+#     web_app_info = types.WebAppInfo(url="https://chatgpt.com/")
+#     menu_button = types.MenuButtonWebApp(text="Open Web App", web_app=web_app_info)
+#     await bot.set_chat_menu_button(chat_id=message.chat.id, menu_button=menu_button)
+#     await message.answer("Web app menu button has been set.")
